deploy/requirements.txt
aws-cdk-lib==2.115.0
boto3-stubs==1.20.20
boto3==1.24.85
botocore==1.27.85
cdk-nag==2.7.2
constructs>=10.0.0,<11.0.0
cdklabs.appsync-utils~=0.0.3
injector==0.21.0